3.0 (llamadas- internacionales y nacionales))

+++ PROYECTO_2_PROGRAAVANZADA/src/test/java/utils/Utils.java
@@ -2,7 +2,7 @@ package utils;
 
 public class Utils {
     private static long CONSECUTIVO = 1;
-    public static final double VALOR_MINUTO = 300;
+    public static final long VALOR_MINUTO = 300;
 
     public static long getCONSECUTIVO() {
         return CONSECUTIVO;
